+++ glue/plugins/tools/pv_slicer/qt/__init__.py
@@ -5,3 +5,4 @@ def setup():
     from glue.viewers.image.qt import ImageViewer
     from glue.plugins.tools.pv_slicer.qt import PVSlicerMode  # noqa
     ImageViewer.tools.append('slice')
+    ImageViewer.tools.append('pv:crosshair')