+++ CHANGES.md
@@ -10,6 +10,8 @@ v1.3.0 (unreleased)
 * Remove bundled version of pvextractor and include it as a proper
   dependency. [#2252]
 
+* Remove ``StandaloneImageViewer`` class. [#1663]
+
 v1.2.4 (2022-01-27)
 -------------------